A0-613 Fix rebase mixup.
